>rename longjump + fly + glide >hycraft crash disabler
