+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/misc/AntiVanish.kt
@@ -37,7 +37,7 @@ class AntiVanish : Module() {
 
     private fun vanish(){
         if((System.currentTimeMillis()-lastNotify)>5000){
-            LiquidBounce.hud.addNotification(Notification("Found a vanished entity!", player.name + " is vanished!!!", NotifyType.WARNING, 4000, 500))
+            LiquidBounce.hud.addNotification(Notification("Found a vanished entity!", player.name + " is vanished", NotifyType.WARNING, 4000, 500))
 
         }
         lastNotify=System.currentTimeMillis()
