+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/movement/flys/hypixel/HypixelFly.kt
@@ -1,22 +1,32 @@
-package net.ccbluex.liquidbounce.features.module.modules.movement.flys.other
+package net.ccbluex.liquidbounce.features.module.modules.movement.flys.hypixel
 
 import net.ccbluex.liquidbounce.event.*
 import net.ccbluex.liquidbounce.features.module.modules.movement.flys.FlyMode
+import net.ccbluex.liquidbounce.utils.timer.MSTimer
 import net.ccbluex.liquidbounce.utils.timer.TickTimer
+import net.ccbluex.liquidbounce.value.BoolValue
 import net.ccbluex.liquidbounce.value.FloatValue
+import net.ccbluex.liquidbounce.value.IntegerValue
 import net.minecraft.block.BlockAir
 import net.minecraft.network.play.client.C03PacketPlayer
 import net.minecraft.util.AxisAlignedBB
 
-class FunCraft : FlyMode("FunCraft") {
-    private val timescale = FloatValue("${valuePrefix}Timer", 1f, 0.1f, 10f)
+class HypixelFly : FlyMode("Hypixel") {
+    private val boostValue = BoolValue("${valuePrefix}Boost", true)
+    private val boostDelayValue = IntegerValue("${valuePrefix}BoostDelay", 1200, 0, 2000)
+    private val boostTimerValue = FloatValue("${valuePrefix}BoostTimer", 1f, 0f, 5f)
+
     private val timer = TickTimer()
+    private val flyTimer = MSTimer()
 
     override fun onUpdate(event: UpdateEvent) {
         fly.antiDesync = true
-        mc.timer.timerSpeed = timescale.get()
-        timer.update()
+        val boostDelay: Long = boostDelayValue.get().toLong()
+        if (boostValue.get() && !flyTimer.hasTimePassed(boostDelay)) {
+            mc.timer.timerSpeed = 1f + boostTimerValue.get() * (flyTimer.hasTimeLeft(boostDelay).toFloat() / boostDelay.toFloat())
+        }
 
+        timer.update()
         if (timer.hasTimePassed(2)) {
             mc.thePlayer.setPosition(mc.thePlayer.posX, mc.thePlayer.posY + 1.0E-5, mc.thePlayer.posZ)
             timer.reset()
