+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/movement/flys/matrix/MatrixBoostFly.kt
@@ -9,7 +9,7 @@ import net.minecraft.network.play.server.S08PacketPlayerPosLook
 import kotlin.math.*
 
 
-class MatrixBoost : FlyMode("MatrixBoost") {
+class MatrixBoostFly : FlyMode("MatrixBoost") {
     private val speed = FloatValue("${valuePrefix}Speed", 2.0f, 1.0f, 3.0f)
     private val jumpTimer = FloatValue("${valuePrefix}JumpTimer", 0.1f, 0.1f, 2f)
     private val boostTimer = FloatValue("${valuePrefix}BoostTimer", 1f, 0.5f, 3f)
