+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/movement/flys/other/MinemoraFly.kt
@@ -1,21 +1,14 @@
 package net.ccbluex.liquidbounce.features.module.modules.movement.flys.other
 
-import net.ccbluex.liquidbounce.LiquidBounce
 import net.ccbluex.liquidbounce.event.EventState
-import net.ccbluex.liquidbounce.ui.client.hud.element.elements.Notification
-import net.ccbluex.liquidbounce.ui.client.hud.element.elements.NotifyType
 import net.ccbluex.liquidbounce.event.*
 import net.ccbluex.liquidbounce.utils.MovementUtils
-import net.ccbluex.liquidbounce.utils.ClientUtils
 import net.ccbluex.liquidbounce.features.module.modules.movement.flys.FlyMode
-import net.minecraft.network.NetHandlerPlayServer
 import net.minecraft.network.Packet
 import net.minecraft.network.play.INetHandlerPlayServer
 import net.minecraft.network.play.client.*
 import net.minecraft.network.play.client.C03PacketPlayer.*
-import java.util.*
 import java.util.concurrent.LinkedBlockingQueue
-import net.minecraft.client.settings.GameSettings
 
 class MinemoraFly : FlyMode("Minemora") {
