+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/movement/flys/vulcan/Vulcan2Fly.kt
@@ -1,4 +1,4 @@
-package net.ccbluex.liquidbounce.features.module.modules.movement.flys.verus
+package net.ccbluex.liquidbounce.features.module.modules.movement.flys.vulcan
 
 import net.ccbluex.liquidbounce.event.PacketEvent
 import net.ccbluex.liquidbounce.event.UpdateEvent
