+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/movement/glides/vulcan/Vulcan263Glide.kt
@@ -3,7 +3,7 @@ package net.ccbluex.liquidbounce.features.module.modules.movement.glides.vulcan
 import net.ccbluex.liquidbounce.event.UpdateEvent
 import net.ccbluex.liquidbounce.features.module.modules.movement.glides.GlideMode
 
-class Vulcan263 : GlideMode("Vulcan2.6.3") {
+class Vulcan263Glide : GlideMode("Vulcan2.6.3") {
     private var ticks = 0
     override fun onUpdate(event: UpdateEvent) {
         if(mc.thePlayer.motionY<= -0.10) {
