--- src/main/java/net/ccbluex/liquidbounce/features/module/modules/movement/longjumps/aac/AACv2.kt
@@ -1,14 +1,0 @@
-package net.ccbluex.liquidbounce.features.module.modules.movement.longjumps.aac
-
-import net.ccbluex.liquidbounce.event.UpdateEvent
-import net.ccbluex.liquidbounce.features.module.modules.movement.longjumps.LongJumpMode
-import net.ccbluex.liquidbounce.utils.MovementUtils
-
-class AACv2 : LongJumpMode("AACv2") {
-    override fun onUpdate(event: UpdateEvent) {
-        mc.thePlayer.jumpMovementFactor = 0.09f
-        mc.thePlayer.motionY += 0.0132099999999999999999999999999
-        mc.thePlayer.jumpMovementFactor = 0.08f
-        MovementUtils.strafe()
-    }
-}
