+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/movement/longjumps/other/HYT4V4Longjump.kt
@@ -4,7 +4,7 @@ import net.ccbluex.liquidbounce.event.UpdateEvent
 import net.ccbluex.liquidbounce.features.module.modules.movement.longjumps.LongJumpMode
 import net.ccbluex.liquidbounce.utils.MovementUtils
 
-class HYT4V4 : LongJumpMode("HYT4V4") {
+class HYT4V4Longjump : LongJumpMode("HYT4V4") {
     override fun onUpdate(event: UpdateEvent) {
         mc.thePlayer.motionY += 0.031470000997
         MovementUtils.strafe(MovementUtils.getSpeed() * 1.0114514f)
